+++ AppCore/Models/TradingCard.py
@@ -1,27 +1,26 @@
 from typing import Any, Dict
+from typing import Optional
 
-
+# Should only be subclassed
 class TradingCard:
     def __init__(self,
                  name: str, 
                  set: str, 
                  type: str,
                  number: str,
-                 double_sided: bool,
-                 json: Dict[str, Any], 
+                 json: Dict[str, Any],
                  metadata: Dict[str, Any] = {}):
         super().__init__()
         self.name: str = name
         self.set: str = set
         self.type: str = type
         self.number: str = number
-        self.double_sided: bool = double_sided
         self.json = json
         self.metadata = metadata
         
     @property
     def is_flippable(self) -> bool:
-        return self.double_sided
+        return self.front_art_url is not None and self.back_art_url is not None
     
     @property
     def friendly_display_name_short(self) -> str:
@@ -34,3 +33,11 @@ class TradingCard:
     @property
     def friendly_display_name_detailed(self) -> str:
         return f'[{self.set+self.number}] {self.name} - {self.type}'
+    
+    @property
+    def front_art_url(self) -> Optional[str]:
+        return NotImplemented
+    
+    @property
+    def back_art_url(self) -> Optional[str]:
+        return NotImplemented