+++ AppCore/Models/__init__.py
@@ -2,3 +2,5 @@ from .LocalCardResource import *
 from .TradingCard import *
 from .SearchConfiguration import *
 from .DeploymentCardResource import *
+from .CardType import *
+from .CardAspect import *