+++ AppUI/Clients/CardImageSource/__init__.py
@@ -1,4 +1,0 @@
-from .SWUDBImageSource import *
-from .SWUDBAPIImageSource import *
-from .CardImageSourceProvider import *
-from .CustomLocalImageSource import *