+++ AppUI/Clients/SWUCardSearchConfiguration.py
@@ -1,10 +1,9 @@
 
 from AppCore.Models import SearchConfiguration
+from AppCore.Models.CardType import CardType
 
-from .CardType import CardType
 
-
-class SWUDBAPISearchConfiguration(SearchConfiguration):
+class SWUCardSearchConfiguration(SearchConfiguration):
     def __init__(self):
         super().__init__()
         self.card_type = CardType.UNSPECIFIED
@@ -12,7 +11,7 @@ class SWUDBAPISearchConfiguration(SearchConfiguration):
     @classmethod
     def from_search_configuration(cls, search_configuration: SearchConfiguration):
         obj = cls.__new__(cls)
-        super(SWUDBAPISearchConfiguration, obj).__init__()
+        super(SWUCardSearchConfiguration, obj).__init__()
         obj.card_name = search_configuration.card_name
         obj.metadata = search_configuration.metadata
         return obj