+++ AppUI/Clients/starwarsunlimited_com/ImageSource.py
@@ -1,23 +1,30 @@
 from AppCore.Resource.CardImageSourceProtocol import RemoteCachedCardImageSourceProtocol
 from AppCore.Models import TradingCard
 from typing import Optional
-
-class SWUDBImageSource(RemoteCachedCardImageSourceProtocol):
+from AppCore.Config import ConfigurationManager
+class ImageSource(RemoteCachedCardImageSourceProtocol):
+    
+    def __init__(self, configuration_manager: ConfigurationManager):
+        self._configuration_manager = configuration_manager
+        
+    @property
+    def configuration_manager(self) -> ConfigurationManager:
+        return self._configuration_manager
     
     @property
     def _site_source_url(self) -> str:
-        return 'https://www.swudb.com/'
+        return 'https://www.starwarsunlimited.com/'
     
     @property
     def site_source_identifier(self) -> str:
-        return 'www.swudb.com'
+        return 'www.starwarsunlimited.com'
     
     @property
     def source_label_display(self) -> str:
         return f'<a href="{self._site_source_url}">{self.site_source_identifier}</a>'
     
     def front_art_url(self, trading_card: TradingCard) -> Optional[str]:
-        return f'https://swudb.com/cards/{trading_card.set}/{trading_card.number}.png'
+        return trading_card.front_art_url
         
     def back_art_url(self, trading_card: TradingCard) -> Optional[str]:
-        return f'https://swudb.com/cards/{trading_card.set}/{trading_card.number}-portrait.png'
+        return trading_card.back_art_url